move subquery segment
